requirements.txt
Flask
mysql-connector-python
flask-bcrypt
Flask-Uploads
